+++ src/pf_arg_x.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   pf_arg_x.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: psentilh <psentilh@student.42.fr>          +#+  +:+       +#+        */
+/*   By: pommedepin <pommedepin@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/03/12 14:50:06 by benjamintle       #+#    #+#             */
-/*   Updated: 2019/03/29 18:20:52 by psentilh         ###   ########.fr       */
+/*   Updated: 2019/03/30 12:48:48 by pommedepin       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ char	*ft_flag_x(va_list list, t_flag flag)
 		if (ft_strcmps(flag.modif, "z") == 0)
 			flag.to_print = ft_ulltoa_base(va_arg(list, unsigned long long), 16, _XMIN_);
 		if (ft_strcmps(flag.modif, "h") == 0)
-			flag.to_print = ft_ulltoa_base(va_arg(list, unsigned int), 16, _XMIN_);
+			flag.to_print = ft_ulltoa_base((unsigned short)va_arg(list, unsigned int), 16, _XMIN_);
 		if (ft_strcmps(flag.modif, "l") == 0)
 			flag.to_print = ft_ulltoa_base(va_arg(list, unsigned long), 16, _XMIN_);
 		if (ft_strcmps(flag.modif, "hh") == 0)
